# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DatasetGenerationRequest": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "examples_per_chunk": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Examples Per Chunk",
            "type": "integer"
          },
          "include_noisy": {
            "default": false,
            "title": "Include Noisy",
            "type": "boolean"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 500.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 25,
            "title": "Limit"
          },
          "min_quality": {
            "default": 0.6,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Quality",
            "type": "number"
          }
        },
        "title": "DatasetGenerationRequest",
        "type": "object"
      },
...
      "DentistVerificationDecision": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "DentistVerificationDecision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Dental AI Chatbot",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/dataset/download": {
      "get": {
        "operationId": "download_dataset_review_csv_api_admin_dataset_download_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Dataset Review Csv",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/dataset/generate": {
      "post": {
        "operationId": "generate_dataset_api_admin_dataset_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetGenerationStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Dataset",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/dataset/status": {
      "get": {
        "operationId": "dataset_generation_status_api_admin_dataset_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetGenerationStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dataset Generation Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/dentist-requests": {
      "get": {
        "description": "Phase 8 (docs/PRODUCT_BENCHMARK.md finding #1): the registration UI has always\npromised \"admin verification required\" for a dentist account - this is the first\nplace an admin can actually see who's asked for one. Oldest request first, so a\nrequest doesn't sit unreviewed just because newer ones keep arriving above it.",
        "operationId": "list_dentist_verification_requests_api_admin_dentist_requests_get",
        "parameters": [
          {
            "description": "pending, approved, rejected, or all",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "description": "pending, approved, rejected, or all",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DentistVerificationRequestRead"
                  },
                  "title": "Response List Dentist Verification Requests Api Admin Dentist Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dentist Verification Requests",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/dentist-requests/{user_id}/approve": {
      "post": {
        "operationId": "approve_dentist_verification_api_admin_dentist_requests__user_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DentistVerificationDecision",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DentistVerificationRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Dentist Verification",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/dentist-requests/{user_id}/reject": {
      "post": {
        "operationId": "reject_dentist_verification_api_admin_dentist_requests__user_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DentistVerificationDecision",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DentistVerificationRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Dentist Verification",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/documents": {
      "get": {
        "operationId": "list_documents_api_admin_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  },
                  "title": "Response List Documents Api Admin Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "upload_document_api_admin_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_admin_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Document",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_admin_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/documents/{document_id}/logs": {
      "get": {
        "operationId": "list_document_ingestion_logs_api_admin_documents__document_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentIngestionLogRead"
                  },
                  "title": "Response List Document Ingestion Logs Api Admin Documents  Document Id  Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Document Ingestion Logs",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}